+++ swarm-test/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Default to `tokio` runtime.
   See [PR 6024](https://github.com/libp2p/rust-libp2p/pull/6024).
+- Remove `async_std` runtime support with `Swarm::new_ephemeral`.
+  Use `Swarm::new_ephemeral_tokio` instead.
+  See [PR 6064](https://github.com/libp2p/rust-libp2p/pull/6064)
 
 - Removed `async-std` support from TCP transport.
   See [PR 5955](https://github.com/libp2p/rust-libp2p/pull/5955).